+++ CHANGES.rst
@@ -5,6 +5,8 @@ grafanimate changelog
 
 in progress
 ===========
+- Add parameters ``--panel-id``, ``--header-layout`` and ``--datetime-format``
+- Refactor some parts of the machinery
 
 
 2018-12-27 0.3.0